move disk space decorator to core
